Protect pages.

+++ app/jobs/unread_limiter.rb
@@ -10,6 +10,7 @@ class UnreadLimiter
 
     subscriptions = Subscription.where(feed_id: feed_id).where(active: true)
     subscriptions.each do |subscription|
+      next if subscription.protected?
       ids = UnreadEntry.where(user_id: subscription.user_id, feed_id: subscription.feed_id).order(published: :desc).offset(entry_limit).pluck(:id)
       ids = ids.last(ids.size * 0.05)
       UnreadEntry.where(id: ids).delete_all
